requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
attrs=19.1.0=py_0
backcall=0.1.0=py37_0
blas=1.0=mkl
bleach=3.1.0=py_0
ca-certificates=2020.1.1=0
certifi=2019.11.28=py37_1
cffi=1.14.0=py37h7a1dbc1_0
chardet=3.0.4=py37_1003
cloudpickle=1.3.0=py_0
cmake=3.16.3=pypi_0
colorama=0.4.3=py_0
configobj=5.0.6=py37_1
constantly=15.1.0=pyh91ea838_0
cudatoolkit=10.1.243=h74a9793_0
cycler=0.10.0=py37_0
dask-core=2.12.0=py_0
decorator=4.4.0=py_0
defusedxml=0.6.0=py_0
dill=0.3.1.1=py37_0
entrypoints=0.3=py37_0
freetype=2.9.1=ha9979f8_1
future=0.17.1=py37_0
google-pasta=0.1.7=py_0
h5py=2.10.0=py37h5e291fa_0
hdf5=1.10.4=h7ebc959_0
icc_rt=2019.0.0=h0cc432a_1
icu=58.2=ha66f8fd_1
imageio=2.8.0=py_0
inplace-abn=1.0.11=pypi_0
intel-openmp=2020.0=166
ipykernel=5.1.4=py37h39e3cac_0
ipython=7.13.0=py37h5ca1d4c_0
ipython_genutils=0.2.0=py37_0
ipywidgets=7.5.1=py_0
jedi=0.15.1=py37_0
jinja2=2.11.1=py_0
joblib=0.14.1=py_0
jpeg=9b=hb83a4c4_2
jsonschema=3.0.2=py37_0
jupyter=1.0.0=py37_7
jupyter_client=6.0.0=py_0
jupyter_console=6.1.0=py_0
jupyter_core=4.6.1=py37_0
keras-applications=1.0.8=py_0
keras-preprocessing=1.1.0=py_1
kiwisolver=1.1.0=py37ha925a31_0
libpng=1.6.37=h2a8f88b_0
libsodium=1.0.16=h9d3ae62_0
libtiff=4.1.0=h56a325e_0
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
markdown=3.1.1=py37_0
markupsafe=1.1.1=py37he774522_0
matplotlib=3.1.1=py37hc8f65d3_0
mistune=0.8.4=py37he774522_0
mkl=2020.0=166
mkl-service=2.3.0=py37hb782905_0
mkl_fft=1.0.15=py37h14836fe_0
mkl_random=1.1.0=py37h675688f_0
msys2-conda-epoch=20160418=1
multicoretsne=0.1=pypi_0
nbconvert=5.6.1=py37_0
nbformat=4.4.0=py37_0
networkx=2.4=py_0
ninja=1.9.0=py37h74a9793_0
nose=1.3.7=py37_2
notebook=6.0.3=py37_0
numpy=1.18.1=py37h93ca92e_0
numpy-base=1.18.1=py37hc3f5095_1
olefile=0.46=py_0
opencv-python=4.2.0.34=pypi_0
openssl=1.1.1f=he774522_0
pandas=1.0.3=py37h47e9c7a_0
pandoc=2.2.3.2=0
pandocfilters=1.4.2=py37_1
parso=0.5.1=py_0
pexpect=4.7.0=py37_0
pickleshare=0.7.5=py37_0
pillow=7.0.0=py37hcc1f983_0
pip=20.0.2=py37_1
prometheus_client=0.7.1=py_0
prompt_toolkit=3.0.3=py_0
protobuf=3.11.3=pypi_0
pycparser=2.20=py_0
pygments=2.4.2=py_0
pyparsing=2.4.2=py_0
pyqt=5.9.2=py37h6538335_2
pyreadline=2.1=py37_1
pyrsistent=0.15.4=py37he774522_0
pyserial=3.4=py37_0
python=3.7.6=h60c2a47_2
python-dateutil=2.8.1=py_0
pytorch=1.4.0=py3.7_cuda101_cudnn7_0
pytz=2019.3=py_0
pywavelets=1.0.3=py37h8c2d366_1
pywin32=227=py37he774522_1
pywinpty=0.5.7=py37_0
pyzmq=18.1.0=py37ha925a31_0
qt=5.9.7=vc14h73c81de_0
qtconsole=4.5.5=py_0
scikit-image=0.14.0=py37h6538335_1
scikit-learn=0.22.1=py37h6288b17_0
scipy=1.3.1=py37h29ff71c_0
seaborn=0.10.0=py_0
send2trash=1.5.0=py37_0
setuptools=46.0.0=py37_0
simplegeneric=0.8.1=py37_2
sip=4.19.8=py37h6538335_0
six=1.12.0=py37_0
sqlite=3.31.1=he774522_0
tensorboardx=2.0=pypi_0
termcolor=1.1.0=py37_1
terminado=0.8.2=py37_0
testpath=0.4.4=py_0
tk=8.6.8=hfa6e2cd_0
toolz=0.10.0=py_0
torchvision=0.5.0=py37_cu101
tornado=6.0.3=py37he774522_3
traitlets=4.3.3=py37_0
vc=14.1=h0510ff6_4
vs2015_runtime=14.16.27012=hf0eaf9b_1
wcwidth=0.1.7=py37_0
webencodings=0.5.1=py37_1
wheel=0.34.2=py37_0
widgetsnbextension=3.5.1=py37_0
wincertstore=0.2=py37_0
winpty=0.4.3=4
wrapt=1.11.2=py37he774522_0
xz=5.2.4=h2fa13f4_4
zeromq=4.3.1=h33f27b4_3
zipp=2.2.0=py_0
zlib=1.2.11=h62dcd97_3
zstd=1.3.7=h508b16e_0
